requirements.txt
Flask
web3
python-dotenv
google-generativeai
requests
paddleocr
paddlepaddle
pytesseract
Pillow
opencv-python
numpy
